backend/apps/common/serializers.py
from rest_framework import serializers

from .models import SiteSettings


class SiteSettingsSerializer(serializers.ModelSerializer):
    class Meta:
        model = SiteSettings
        fields = ("restore_throttle_rate", "manual_backup_throttle_rate")
