backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-filter>=24.0
django-cors-headers>=4.6.0
psycopg[binary]>=3.1.18
PyMySQL>=1.1.1
celery>=5.4.0
redis>=5.0.0
python-dotenv>=1.0.1
cryptography>=42.0.0
paramiko>=3.4.0
